Refactor LocalBuild so it's not static
